Merge Compacts into GHC
Summary:
Compacts are a new data structure that was illustrated at ICFP 2015.
It is an immutable NF data structure with no outgoing pointers.
This structure can be serialized quickly and can be GCed without
ever traversing it (it is a sort of "older than oldest generation",
but it's not implemented as such because it would be a mess)

Test Plan:
validate
(new test cases included)

Reviewers: simonmar, austin, ezyang, hvr, bgamari

Subscribers: thomie, erikd

Differential Revision: https://phabricator.haskell.org/D1264